Rename controllers folder to controller. Fix babel configuration to support import and export default statements in the backend. Use import and export default statments in routes and index server. Update controller tests to import from correct folder.
