+++ package.json
@@ -8,8 +8,8 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "server-dev": "nodemon src/server/index.js",
-    "start": "node src/server/index.js",
+    "server-dev": "babel-node -x \".js\" src/server/index.js",
+    "start": "babel-node -x \".js\" src/server/index.js",
     "test": "mocha test/**/*.test.js --require ./test/dom.js --require esm --require ignore-styles --require @babel/register ./test --exit || true",
     "seed": "node ./seed/index.js",
     "react-dev": "webpack -d --watch",
@@ -30,8 +30,9 @@
     "winston": "^3.3.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
+    "@babel/core": "^7.12.3",
+    "@babel/node": "^7.12.6",
+    "@babel/preset-env": "^7.12.1",
     "@babel/preset-react": "^7.10.4",
     "@babel/register": "^7.12.1",
     "babel-loader": "^8.1.0",
@@ -44,7 +45,7 @@
     "ignore-styles": "^5.0.1",
     "jsdom": "^16.4.0",
     "mocha": "^8.1.3",
-    "nodemon": "^2.0.4",
+    "nodemon": "^2.0.6",
     "sinon": "^9.2.0",
     "style-loader": "^2.0.0",
     "supertest": "^6.0.1",
