+++ src/server/index.js
@@ -1,8 +1,9 @@
-const app = require("./routes");
+import app from "./routes";
+
 const port = process.env.PORT || 3004;
 
 app.listen(port, () => {
   console.log(`Listening on port ${port}`);
 });
 
-module.exports = app;
+export default app;
